--- src/main/java/me/xhsun/guildwars2wrapper/requests/v2/account/Achievements.java
@@ -1,57 +1,0 @@
-package me.xhsun.guildwars2wrapper.requests.v2.account;
-
-import java.io.IOException;
-import java.util.List;
-import me.xhsun.guildwars2wrapper.error.ErrorCode;
-import me.xhsun.guildwars2wrapper.error.GuildWars2Exception;
-import me.xhsun.guildwars2wrapper.model.v2.account.AchievementProgression;
-import me.xhsun.guildwars2wrapper.requests.GuildWars2API;
-import me.xhsun.guildwars2wrapper.requests.Request;
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
-public class Achievements extends Request {
-
-	protected Achievements(GuildWars2API gw2API) {
-		super(gw2API);
-	}
-
-	//ASYNC
-	/**
-	 * For more info on Account API go <a href="https://wiki.guildwars2.com/wiki/API:2/account/achievements">here</a><br/>
-	 * Give user the access to {@link Callback#onResponse(Call, Response)} and {@link Callback#onFailure(Call, Throwable)} methods for custom interactions
-	 *
-	 * @param API      API key
-	 * @param callback callback that is going to be used for {@link Call#enqueue(Callback)}
-	 * @throws GuildWars2Exception  invalid API key
-	 * @throws NullPointerException if given {@link Callback} is empty
-	 * @see AchievementProgression Account achievement info
-	 */
-	public void getAchievementProgression(String API, Callback<List<AchievementProgression>> callback) throws GuildWars2Exception, NullPointerException {
-		isParamValid(new ParamChecker(ParamType.API, API));
-		gw2API.getAchievementProgression(API).enqueue(callback);
-	}
-
-	//SYNC
-	/**
-	 * For more info on Account API go <a href="https://wiki.guildwars2.com/wiki/API:2/account/achievements">here</a><br/>
-	 * Get an account's progress towards all their achievements.
-	 *
-	 * @param API API key
-	 * @return list of account achievement info
-	 * @throws GuildWars2Exception see {@link ErrorCode} for detail
-	 * @see AchievementProgression account achievement info
-	 */
-	public List<AchievementProgression> getAchievementProgression(String API) throws GuildWars2Exception {
-		isParamValid(new ParamChecker(ParamType.API, API));
-		try {
-			Response<List<AchievementProgression>> response = gw2API.getAchievementProgression(API).execute();
-			if (!response.isSuccessful()) throwError(response.code(), response.errorBody());
-			return response.body();
-		} catch (IOException e) {
-			throw new GuildWars2Exception(ErrorCode.Network, "Network Error: " + e.getMessage());
-		}
-	}
-
-}
