--- src/main/java/me/xhsun/guildwars2wrapper/requests/v2/account/Dungeons.java
@@ -1,52 +1,0 @@
-package me.xhsun.guildwars2wrapper.requests.v2.account;
-
-import java.io.IOException;
-import java.util.List;
-import me.xhsun.guildwars2wrapper.error.ErrorCode;
-import me.xhsun.guildwars2wrapper.error.GuildWars2Exception;
-import me.xhsun.guildwars2wrapper.requests.GuildWars2API;
-import me.xhsun.guildwars2wrapper.requests.Request;
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
-public class Dungeons extends Request {
-
-	protected Dungeons(GuildWars2API gw2API) {
-		super(gw2API);
-	}
-
-	//ASYNC
-	/**
-	 * For more info on Dungeon progression API go <a href="https://wiki.guildwars2.com/wiki/API:2/account/dungeons">here</a><br/>
-	 * Give user the access to {@link Callback#onResponse(Call, Response)} and {@link Callback#onFailure(Call, Throwable)} methods for custom interactions
-	 *
-	 * @param API      API key
-	 * @param callback callback that is going to be used for {@link Call#enqueue(Callback)}
-	 * @throws GuildWars2Exception  invalid API key
-	 * @throws NullPointerException if given {@link Callback} is empty
-	 */
-	public void getDailyDungeonProgression(String API, Callback<List<String>> callback) throws GuildWars2Exception, NullPointerException {
-		isParamValid(new ParamChecker(ParamType.API, API));
-		gw2API.getDailyDungeonProgression(API).enqueue(callback);
-	}
-
-	//SYNC
-	/**
-	 * For more info on Dungeon progression API go <a href="https://wiki.guildwars2.com/wiki/API:2/account/dungeons">here</a><br/>
-	 *
-	 * @param API API key
-	 * @return an array of strings representing dungeon path names completed since daily dungeon reset
-	 * @throws GuildWars2Exception see {@link ErrorCode} for detail
-	 */
-	public List<String> getDailyDungeonProgression(String API) throws GuildWars2Exception {
-		isParamValid(new ParamChecker(ParamType.API, API));
-		try {
-			Response<List<String>> response = gw2API.getDailyDungeonProgression(API).execute();
-			if (!response.isSuccessful()) throwError(response.code(), response.errorBody());
-			return response.body();
-		} catch (IOException e) {
-			throw new GuildWars2Exception(ErrorCode.Network, "Network Error: " + e.getMessage());
-		}
-	}
-}
