--- src/main/java/me/xhsun/guildwars2wrapper/requests/v2/account/Dyes.java
@@ -1,56 +1,0 @@
-package me.xhsun.guildwars2wrapper.requests.v2.account;
-
-import java.io.IOException;
-import java.util.List;
-import me.xhsun.guildwars2wrapper.error.ErrorCode;
-import me.xhsun.guildwars2wrapper.error.GuildWars2Exception;
-import me.xhsun.guildwars2wrapper.model.v2.Color;
-import me.xhsun.guildwars2wrapper.requests.GuildWars2API;
-import me.xhsun.guildwars2wrapper.requests.Request;
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
-public class Dyes extends Request {
-
-	protected Dyes(GuildWars2API gw2API) {
-		super(gw2API);
-	}
-
-	//ASYNC
-	/**
-	 * For more info on Account dyes API go <a href="https://wiki.guildwars2.com/wiki/API:2/account/dyes">here</a><br/>
-	 * Get list of unlocked dyes ids linked to given API key
-	 *
-	 * @param API      API key
-	 * @param callback callback that is going to be used for {@link Call#enqueue(Callback)}
-	 * @throws GuildWars2Exception  invalid API key
-	 * @throws NullPointerException if given {@link Callback} is empty
-	 * @see Color color info
-	 */
-	public void getUnlockedDyes(String API, Callback<List<Integer>> callback) throws GuildWars2Exception, NullPointerException {
-		isParamValid(new ParamChecker(ParamType.API, API));
-		gw2API.getUnlockedDyes(API).enqueue(callback);
-	}
-
-	//SYNC
-	/**
-	 * For more info on Account dyes API go <a href="https://wiki.guildwars2.com/wiki/API:2/account/dyes">here</a><br/>
-	 * Get list of unlocked dyes ids linked to given API key
-	 *
-	 * @param API API key
-	 * @return list of color ids
-	 * @throws GuildWars2Exception see {@link ErrorCode} for detail
-	 * @see Color
-	 */
-	public List<Integer> getUnlockedDyes(String API) throws GuildWars2Exception {
-		isParamValid(new ParamChecker(ParamType.API, API));
-		try {
-			Response<List<Integer>> response = gw2API.getUnlockedDyes(API).execute();
-			if (!response.isSuccessful()) throwError(response.code(), response.errorBody());
-			return response.body();
-		} catch (IOException e) {
-			throw new GuildWars2Exception(ErrorCode.Network, "Network Error: " + e.getMessage());
-		}
-	}
-}
