--- src/main/java/me/xhsun/guildwars2wrapper/requests/v2/account/Finishers.java
@@ -1,56 +1,0 @@
-package me.xhsun.guildwars2wrapper.requests.v2.account;
-
-import java.io.IOException;
-import java.util.List;
-import me.xhsun.guildwars2wrapper.error.ErrorCode;
-import me.xhsun.guildwars2wrapper.error.GuildWars2Exception;
-import me.xhsun.guildwars2wrapper.model.v2.account.UnlockedFinisher;
-import me.xhsun.guildwars2wrapper.requests.GuildWars2API;
-import me.xhsun.guildwars2wrapper.requests.Request;
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
-public class Finishers extends Request {
-
-	protected Finishers(GuildWars2API gw2API) {
-		super(gw2API);
-	}
-
-	//ASYNC
-	/**
-	 * For more info on Account finishers API go <a href="https://wiki.guildwars2.com/wiki/API:2/account/finishers">here</a><br/>
-	 * Get list of unlocked finishers linked to given API key
-	 *
-	 * @param API      API key
-	 * @param callback callback that is going to be used for {@link Call#enqueue(Callback)}
-	 * @throws GuildWars2Exception  invalid API key
-	 * @throws NullPointerException if given {@link Callback} is empty
-	 * @see UnlockedFinisher unlocked finisher info
-	 */
-	public void getUnlockedFinishers(String API, Callback<List<UnlockedFinisher>> callback) throws GuildWars2Exception, NullPointerException {
-		isParamValid(new ParamChecker(ParamType.API, API));
-		gw2API.getUnlockedFinishers(API).enqueue(callback);
-	}
-
-	//SYNC
-	/**
-	 * For more info on Account finishers API go <a href="https://wiki.guildwars2.com/wiki/API:2/account/finishers">here</a><br/>
-	 * Get list of unlocked finishers linked to given API key
-	 *
-	 * @param API API key
-	 * @return list of unlocked finisher info
-	 * @throws GuildWars2Exception see {@link ErrorCode} for detail
-	 * @see UnlockedFinisher unlocked finisher info
-	 */
-	public List<UnlockedFinisher> getUnlockedFinishers(String API) throws GuildWars2Exception {
-		isParamValid(new ParamChecker(ParamType.API, API));
-		try {
-			Response<List<UnlockedFinisher>> response = gw2API.getUnlockedFinishers(API).execute();
-			if (!response.isSuccessful()) throwError(response.code(), response.errorBody());
-			return response.body();
-		} catch (IOException e) {
-			throw new GuildWars2Exception(ErrorCode.Network, "Network Error: " + e.getMessage());
-		}
-	}
-}
