--- src/main/java/me/xhsun/guildwars2wrapper/requests/v2/account/Mailcarriers.java
@@ -1,33 +1,0 @@
-package me.xhsun.guildwars2wrapper.requests.v2.account;
-
-import java.util.List;
-import me.xhsun.guildwars2wrapper.error.GuildWars2Exception;
-import me.xhsun.guildwars2wrapper.requests.GuildWars2API;
-import me.xhsun.guildwars2wrapper.requests.Request;
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
-public class Mailcarriers extends Request {
-
-	protected Mailcarriers(GuildWars2API gw2API) {
-		super(gw2API);
-	}
-
-	//ASYNC
-	/**
-	 * For more info on account mail carrier API go <a href="https://wiki.guildwars2.com/wiki/API:2/account/mailcarriers">here</a><br/>
-	 * Give user the access to {@link Callback#onResponse(Call, Response)} and {@link Callback#onFailure(Call, Throwable)} methods for custom interactions
-	 *
-	 * @param API      API key
-	 * @param callback callback that is going to be used for {@link Call#enqueue(Callback)}
-	 * @throws GuildWars2Exception  invalid API key
-	 * @throws NullPointerException if given {@link Callback} is empty
-	 */
-	public void getUnlockedMailCarriers(String API, Callback<List<Integer>> callback) throws GuildWars2Exception, NullPointerException {
-		isParamValid(new ParamChecker(ParamType.API, API));
-		gw2API.getUnlockedMailCarriers(API).enqueue(callback);
-	}
-
-	//SYNC
-}
